# specs/002-make-a-call/contracts/api-contracts.yaml
openapi: 3.1.0
info:
  title: promptalicious API
  description: API for LLM prompt execution and diagnostics
  version: 0.2.0
servers:
  - url: http://localhost:3000/api
    description: Local development server
paths:
...
  /config:
    get:
      operationId: getConfiguration
      summary: Get current LLM configuration
      description: |
        Returns the current LLM configuration (model selection, endpoint).
        API key is never returned for security reasons.

        Functional Requirements: FR-020, FR-023
      tags:
        - Configuration
      responses:
        '200':
          description: Configuration retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigurationResponse'
              examples:
                defaultConfig:
                  summary: Default configuration
                  value:
                    config:
                      id: 1
                      selectedModel: "gpt-4o-mini"
                      providerEndpoint: null
                      createdAt: "2025-11-04T09:00:00.000Z"
                      updatedAt: "2025-11-04T09:00:00.000Z"
                    availableModels:
                      - "gpt-4o-mini"

    put:
      operationId: updateConfiguration
      summary: Update LLM configuration
      description: |
        Updates the LLM configuration (model selection, API key, endpoint).
        Automatically validates credentials via test call before saving.

        Functional Requirements: FR-020, FR-021, FR-022, FR-023, FR-024, FR-024a
      tags:
        - Configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateConfigurationRequest'
            examples:
              updateApiKey:
                summary: Update API key
                value:
                  apiKey: "sk-proj-newkey123456789"
              updateModel:
                summary: Change model (future expansion)
                value:
                  selectedModel: "gpt-4o-mini"
              fullUpdate:
                summary: Update multiple fields
                value:
                  selectedModel: "gpt-4o-mini"
                  apiKey: "sk-proj-newkey123456789"
                  providerEndpoint: "https://api.openai.com/v1"
      responses:
        '200':
          description: Configuration updated and validated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateConfigurationSuccessResponse'
              examples:
                successfulUpdate:
                  summary: Configuration saved after validation
                  value:
                    config:
                      id: 1
                      selectedModel: "gpt-4o-mini"
                      providerEndpoint: null
                      updatedAt: "2025-11-04T10:40:00.000Z"
                    validationResult:
                      success: true
                      message: "API credentials validated successfully"
        '400':
          description: Validation failed - invalid credentials or configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateConfigurationErrorResponse'
              examples:
                invalidCredentials:
                  summary: Test call failed
                  value:
                    error:
                      errorType: "authentication"
                      errorMessage: "API key validation failed: invalid credentials"
                      additionalContext:
                        testCallFailed: true
  /config/test-connection:
    post:
      operationId: testConnection
      summary: Test LLM API connection
      description: |
        Manually test the connection to the LLM provider using current configuration.
        Does not save any changes - purely for validation.

        Functional Requirements: FR-024a
      tags:
        - Configuration
      responses:
        '200':
          description: Connection test successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestConnectionResponse'
              examples:
                connectionSuccess:
                  summary: Valid credentials
                  value:
                    success: true
                    message: "Successfully connected to OpenAI API with GPT-4o-mini"
        '400':
          description: Connection test failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestConnectionResponse'
              examples:
                connectionFailure:
                  summary: Invalid credentials
                  value:
                    success: false
                    message: "Connection failed: Invalid API key"
                    error:
                      errorType: "authentication"
                      errorCode: "invalid_api_key"
...
components:
  schemas:
...
    ConfigurationResponse:
      type: object
      required:
        - config
        - availableModels
      properties:
        config:
          $ref: '#/components/schemas/LLMConfiguration'
        availableModels:
          type: array
          items:
            type: string
          description: List of available model identifiers
          example: ["gpt-4o-mini"]
    LLMConfiguration:
      type: object
      required:
        - id
        - selectedModel
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          enum: [1]
          description: Always 1 (single-row table)
        selectedModel:
          type: string
          description: Currently selected model
          example: "gpt-4o-mini"
        providerEndpoint:
          type: string
          format: uri
          nullable: true
          description: Custom endpoint URL (optional)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        # Note: apiKey intentionally excluded from response schema for security
    UpdateConfigurationRequest:
      type: object
      properties:
        selectedModel:
          type: string
          description: Model to use (must be in availableModels list)
          example: "gpt-4o-mini"
        apiKey:
          type: string
          description: API key for the provider (will be encrypted)
          minLength: 1
        providerEndpoint:
          type: string
          format: uri
          description: Custom endpoint URL (optional)
      # At least one field must be provided
      minProperties: 1
    UpdateConfigurationSuccessResponse:
      type: object
      required:
        - config
        - validationResult
      properties:
        config:
          $ref: '#/components/schemas/LLMConfiguration'
        validationResult:
          type: object
          required:
            - success
            - message
          properties:
            success:
              type: boolean
              enum: [true]
            message:
              type: string
              example: "API credentials validated successfully"
    UpdateConfigurationErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - errorType
            - errorMessage
          properties:
            errorType:
              type: string
              enum: [authentication, validation]
            errorMessage:
              type: string
            additionalContext:
              type: object
              additionalProperties: true
    TestConnectionResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
        message:
          type: string
        error:
          type: object
          properties:
            errorType:
              type: string
            errorCode:
              type: string
...